Implemented error message handling in the frontend
Created another grid row for the error message,
used a Binding and BooleanToVisibilityConverter to have the error message dynamically appear and disappear, with a fallback value of collapsed.

Had to add <BooleanToVisibilityConverter in the resource dictionary to use it.

Created 2 props in LoginViewModel, IsErrorVisible, for the dynamic appearance of the error box in the LoginView, and ErrorMessage to display the error message in the loginView